Move scene construction into renderer state
