+++ src/heart/renderers/artist/__init__.py
@@ -1,4 +1,2 @@
-from heart.renderers.artist.provider import \
-    ArtistStateProvider as ArtistStateProvider
 from heart.renderers.artist.renderer import ArtistScene as ArtistScene
 from heart.renderers.artist.state import ArtistState as ArtistState
