+++ src/heart/renderers/artist/renderer.py
@@ -1,15 +1,14 @@
 from heart.display.color import Color
 from heart.navigation import MultiScene
 from heart.renderers import BaseRenderer
-from heart.renderers.artist.provider import ArtistStateProvider
+from heart.renderers.artist.state import ArtistState
 from heart.renderers.image import RenderImage
 from heart.renderers.text import TextRendering
 
 
 class ArtistScene(MultiScene):
-    def __init__(self, provider: ArtistStateProvider | None = None) -> None:
-        self._provider = provider or ArtistStateProvider()
-        artist_state = self._provider.build()
+    def __init__(self) -> None:
+        artist_state = ArtistState.build()
         super().__init__(artist_state.scenes)
 
     @staticmethod
