+++ src/heart/renderers/kirby/__init__.py
@@ -1,4 +1,2 @@
-from heart.renderers.kirby.provider import \
-    KirbyStateProvider as KirbyStateProvider
 from heart.renderers.kirby.renderer import KirbyScene as KirbyScene
 from heart.renderers.kirby.state import KirbyState as KirbyState
